requirements.txt
langgraph>=0.2.34
langchain>=0.2.12
langchain-openai>=0.1.20
openai>=1.13.3
pydantic>=2.7.0
tenacity>=8.4.2
numpy>=1.26.4
pypdf>=4.2.0
python-docx>=1.1.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
rich>=13.7.1
